# docs/swagger.yaml
basePath: /
definitions:
  models.User:
    type: object
host: localhost:8000
info:
  contact:
    name: nathan.lu
    url: https://tedmax100.github.io/
  description: book service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Book service
  version: "1.0"
paths:
  /{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: token
        in: header
        name: auth
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Delete User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: token
        in: header
        name: auth
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Update User
      tags:
      - User
  /users:
    get:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get User
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Create User
      tags:
      - User
schemes:
- http,https
swagger: "2.0"